+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/codec/LuceneOptimizedPostingsFormat.java
@@ -44,16 +44,19 @@ import java.util.Iterator;
 @AutoService(PostingsFormat.class)
 public class LuceneOptimizedPostingsFormat extends PostingsFormat {
     PostingsFormat postingsFormat;
-    boolean allowIntegrityCheck = true;
+    private static boolean allowCheckDataIntegrity = true;
 
     public LuceneOptimizedPostingsFormat() {
-        this(new Lucene84PostingsFormat(), true);
+        this(new Lucene84PostingsFormat());
     }
 
-    public LuceneOptimizedPostingsFormat(PostingsFormat postingsFormat, boolean allowIntegrityCheck) {
+    public LuceneOptimizedPostingsFormat(PostingsFormat postingsFormat) {
         super("RL" + postingsFormat.getName());
         this.postingsFormat = postingsFormat;
-        this.allowIntegrityCheck = allowIntegrityCheck;
+    }
+
+    public static void setAllowCheckDataIntegrity(boolean allow) {
+        allowCheckDataIntegrity = allow;
     }
 
     @Override
@@ -64,7 +67,7 @@ public class LuceneOptimizedPostingsFormat extends PostingsFormat {
     @Override
     @SuppressWarnings("PMD.CloseResource")
     public FieldsProducer fieldsProducer(SegmentReadState state) throws IOException {
-        return new LazyFieldsProducer(state, allowIntegrityCheck);
+        return new LazyFieldsProducer(state);
     }
 
     private static class LazyFieldsProducer extends FieldsProducer {
@@ -72,10 +75,8 @@ public class LuceneOptimizedPostingsFormat extends PostingsFormat {
         private Supplier<FieldsProducer> fieldsProducer;
 
         private boolean initialized;
-        private boolean allowIntegrityCheck;
 
-        private LazyFieldsProducer(final SegmentReadState state, boolean allowIntegrityCheck) {
-            this.allowIntegrityCheck = allowIntegrityCheck;
+        private LazyFieldsProducer(final SegmentReadState state) {
             fieldsProducer = Suppliers.memoize(() -> {
                 try {
                     PostingsReaderBase postingsReader = new LuceneOptimizedPostingsReader(state);
@@ -97,7 +98,7 @@ public class LuceneOptimizedPostingsFormat extends PostingsFormat {
 
         @Override
         public void checkIntegrity() throws IOException {
-            if (allowIntegrityCheck) {
+            if (allowCheckDataIntegrity) {
                 fieldsProducer.get().checkIntegrity();
             }
         }