+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/directory/FDBLuceneFunctionalityTest.java
@@ -20,6 +20,7 @@
 
 package com.apple.foundationdb.record.lucene.directory;
 
+import com.apple.foundationdb.record.lucene.codec.LuceneOptimizedPostingsFormat;
 import com.apple.test.Tags;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.document.Document;
@@ -68,7 +69,8 @@ public class FDBLuceneFunctionalityTest extends FDBDirectoryBaseTest {
 
     @Test
     public void givenTermQueryWhenFetchedDocumentThenCorrect() throws Exception {
-        luceneIndex.indexDocument("activity", "running in track", false);
+        LuceneOptimizedPostingsFormat.setAllowCheckDataIntegrity(false);
+        luceneIndex.indexDocument("activity", "running in track");
         luceneIndex.indexDocument("activity", "Cars are running on road");
         Term term = new Term("body", "running");
         Query query = new TermQuery(term);
@@ -89,7 +91,8 @@ public class FDBLuceneFunctionalityTest extends FDBDirectoryBaseTest {
     @Test
     public void givenBooleanQueryWhenFetchedDocumentThenCorrect() throws Exception {
         luceneIndex.indexDocument("Destination", "Las Vegas singapore car");
-        luceneIndex.indexDocument("Commutes in singapore", "Bus Car Bikes", false);
+        LuceneOptimizedPostingsFormat.setAllowCheckDataIntegrity(false);
+        luceneIndex.indexDocument("Commutes in singapore", "Bus Car Bikes");
         Term term1 = new Term("body", "singapore");
         Term term2 = new Term("body", "car");
         TermQuery query1 = new TermQuery(term1);
@@ -110,7 +113,8 @@ public class FDBLuceneFunctionalityTest extends FDBDirectoryBaseTest {
 
     @Test
     public void givenFuzzyQueryWhenFetchedDocumentThenCorrect() throws Exception {
-        luceneIndex.indexDocument("article", "Halloween Festival", false);
+        LuceneOptimizedPostingsFormat.setAllowCheckDataIntegrity(false);
+        luceneIndex.indexDocument("article", "Halloween Festival");
         luceneIndex.indexDocument("decoration", "Decorations for Halloween");
         Term term = new Term("body", "hallowen");
         Query query = new FuzzyQuery(term);
@@ -132,7 +136,8 @@ public class FDBLuceneFunctionalityTest extends FDBDirectoryBaseTest {
     public void givenSortFieldWhenSortedThenCorrect() throws Exception {
         luceneIndex.indexDocument("Ganges", "River in India");
         luceneIndex.indexDocument("Mekong", "This river flows in south Asia");
-        luceneIndex.indexDocument("Amazon", "Rain forest river", false);
+        LuceneOptimizedPostingsFormat.setAllowCheckDataIntegrity(false);
+        luceneIndex.indexDocument("Amazon", "Rain forest river");
         luceneIndex.indexDocument("Rhine", "Belongs to Europe");
         luceneIndex.indexDocument("Nile", "Longest River");
 
@@ -149,6 +154,7 @@ public class FDBLuceneFunctionalityTest extends FDBDirectoryBaseTest {
 
     @Test
     public void whenDocumentDeletedThenCorrect() throws IOException {
+        LuceneOptimizedPostingsFormat.setAllowCheckDataIntegrity(false);
         luceneIndex.indexDocument("Ganges", "River in India");
         luceneIndex.indexDocument("Mekong", "This river flows in south Asia");
         Term term = new Term("title", "ganges");